+++ entrypoints/index.content.ts
@@ -2,6 +2,7 @@ import { initiateRouter, blockPage } from "@/utils/initiator/router"
 
 const watchPattern = new MatchPattern("*://www.nicovideo.jp/watch/*")
 const rankingPattern = new MatchPattern("*://www.nicovideo.jp/ranking*")
+const recommendationPattern = new MatchPattern("*://www.nicovideo.jp/recommendations*")
 
 const searchPatternArray = [
     new MatchPattern("*://www.nicovideo.jp/search/*"),
@@ -12,6 +13,13 @@ const searchPatternArray = [
     new MatchPattern("*://www.nicovideo.jp/mylist_search/*"),
     new MatchPattern("*://www.nicovideo.jp/user_search/*"),
 ] as const
+
+export type catchMatchFor = {
+    watch: boolean
+    ranking: boolean
+    search: boolean
+    recommendations: boolean
+}
 
 export default defineContentScript({
     matches: ["*://www.nicovideo.jp/*"],
@@ -38,10 +46,29 @@ export default defineContentScript({
                 || (isSearch && enableSearchPage)
             ) {
                 initiateRouter(ctx, storage)
-            } else if ((!enableReshogi && isRanking) || (!enableSearchPage && isSearch)) {
-                ctx.addEventListener(window, "wxt:locationchange", ({ newUrl }) => {
-                    if (watchPattern.includes(newUrl) || (rankingPattern.includes(newUrl) && enableReshogi)) window.location.reload()// Promise.allSettled(storagePromises).then(initializeWatch, onError);
+            } else if ((!enableReshogi && isRanking) || (!enableSearchPage && isSearch) || recommendationPattern) {
+                const matchFor = {
+                    watch: true,
+                    ranking: enableReshogi,
+                    search: enableSearchPage,
+                    recommendations: false,
+                }
+                injectScript("/catchTargetPage.js", {
+                    modifyScript(script) {
+                        // MAINスクリプト側から引っかかったことを受け取ったらreload
+                        script.addEventListener("mwReactRouterHit", (event) => {
+                            if (event instanceof CustomEvent) {
+                                console.log(`${event.type}:`, event.detail)
+                            }
+                            window.location.reload()
+                        })
+                        // どのページが有効化されているかをMAINスクリプト側に伝える
+                        script.dataset["matchfor"] = JSON.stringify(matchFor)
+                    },
                 })
+                /* ctx.addEventListener(window, "wxt:locationchange", ({ newUrl }) => {
+                    if (watchPattern.includes(newUrl) || (rankingPattern.includes(newUrl) && enableReshogi)) window.location.reload()// Promise.allSettled(storagePromises).then(initializeWatch, onError);
+                }) */
             }
         })
     },